+++ synapseclient/core/utils.py
@@ -880,3 +880,8 @@ def require_param(param, name):
     if param is None:
         raise ValueError("%s parameter is required." % name)
 
+
+def snake_case(string):
+    """Convert the given string from CamelCase to snake_case"""
+    # https://stackoverflow.com/a/1176023
+    return re.sub(r'(?<!^)(?=[A-Z])', '_', string).lower()